Refactored scripts folder structure, fields access modifiers and some comments
