+++ Assets/Scripts/Player/PlayerProjectileMoving.cs
@@ -4,13 +4,6 @@ public class PlayerProjectileMoving : MonoBehaviour
 {
     public float speed;
 
-    // Start is called before the first frame update
-    void Start()
-    {
-
-    }
-
-    // Update is called once per frame
     void Update()
     {
         transform.Translate(Vector2.up * speed * Time.deltaTime);
