+++ Assets/Scripts/SpawnProjectile.cs
@@ -10,13 +10,6 @@ public class SpawnProjectile : MonoBehaviour
 
     bool shouldShoot = true;
 
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
     void Update()
     {
         HandleShooting();
