+++ Assets/Scripts/Ufo.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f390b2d79f1f2c241aba76dec77e339e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
