+++ Assets/Scripts/Ufo/UfoCollisions.cs
@@ -1,20 +1,19 @@
 ﻿using UnityEngine;
 
-public class UfoProjectileCollisions : MonoBehaviour
+public class UfoCollisions : MonoBehaviour
 {
-    GameManager game;
     SpawnAsteroids spawner;
 
     void Start()
     {
-        game = FindObjectOfType<GameManager>();
-        spawner = FindObjectOfType<SpawnAsteroids>();        
+        spawner = FindObjectOfType<SpawnAsteroids>();
     }
 
     private void OnTriggerEnter2D(Collider2D other)
     {
-        // extract to separate function
-        if (other.gameObject.tag.Contains("Asteroid"))
+        var otherTag = other.gameObject.tag;
+
+        if (otherTag.Contains("Asteroid"))
         {
             if (other.CompareTag("Asteroid_big"))
             {
@@ -27,10 +26,6 @@ public class UfoProjectileCollisions : MonoBehaviour
 
             Destroy(other.gameObject);
             Destroy(gameObject);
-        }
-        else if (other.gameObject.CompareTag("Player"))
-        {
-            game.HandlePlayerDamage(other);
         }
     }
 }
